refactor: Enum 파일 이동
