+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/item/domain/beverage/enums/BeverageSizeOption.java
@@ -1,4 +1,4 @@
-package git_kkalnane.backend.starbucks.item.domain;
+package git_kkalnane.backend.starbucks.item.domain.beverage.enums;
 
 public enum BeverageSizeOption {
     TALL(0),
